Adding KclInternetSocket class as a prototype for an IPv4/IPv6 capable Socket. Client side is working but only if the dynamically allocated addresses are not deleted from the heap, which needs to be fixed of course.

+++ Sources/SocksCore/Types.swift
@@ -115,9 +115,9 @@ extension Protocol: CTypeInt32Convertible {
 extension AddressFamily: CTypeInt32Convertible {
     func toCType() -> Int32 {
         switch self {
-        case .Inet: return AF_INET
-        case .Inet6: return AF_INET6
-        case .UNSPECIFIED : return AF_UNSPEC
+        case .Inet: return Int32(AF_INET)
+        case .Inet6: return Int32(AF_INET6)
+        case .UNSPECIFIED : return Int32(AF_UNSPEC)
         }
     }
 }
